Remove unused argument outputContainer from checkLayout() method
checkLayout() method defined in resources/check-layout-th.js
had a outputContainer argument that was not used.
This patch removes that argument.

On top of that, some tests that were trying to use that argument
have been modified as it was not needed for them.

Last, the copies of check-layout-th.js in WPT CSS test suites
have been updated with this change and another change introduced
in r377261.

Change-Id: Iefb649d811e0fd09106bf3d02225539e481fa713